chore: Update clinical test controller
Refactor the `addTest` function to save the new clinical test and send the response with the created test object. Remove commented out code for generating dummy data.

Also, modify the `getTest` function to send the response using the `sendResponse` function and include a success message along with the fetched clinical tests.

```