chore: Refactor user and department routes
This commit refactors the user and department routes to improve code organization and readability. The user.route.js file is updated to use the createUser, loginUser, and getUser functions from the user.controller.js file. The department.route.js file is updated to include new routes for getting, updating, and deleting departments. The department.controller.js file is also updated to include the deleteDepartment and updateDepartment functions. These changes enhance the functionality and maintainability of the application.