feat: Populate appointment department in FindBy controller
This commit modifies the `FindBy` controller in the `prescription.controller.js` file. It adds a `populate` method to the `Prescription.findById` query, allowing the `appointment` field to be populated with the associated `department` model. This change enhances the functionality of the application by providing access to the department information when retrieving a prescription by ID.