feat: Add deleteAdmin API
This commit adds the `deleteAdmin` API to the `admin.controller.js` file. It allows the deletion of an admin user by their ID. If the admin is found and successfully deleted, a success response is sent. If the admin is not found, a 404 error response is sent. This addition enhances the functionality of the application.