feat: Update updateUser API to set isCompleted flag
This commit updates the `updateUser` API in the `user.controller.js` file to set the `isCompleted` flag based on whether all required fields are filled. If all fields (fullName, phone, district, upazila, address) are provided, the `isCompleted` flag is set to true. Otherwise, a message is generated indicating which fields are missing. This enhancement improves the user update functionality.